Removes the transpose of box in read_xdr
I don't think we need to transpose box, as long as we keep in mind that
the array dimensions are reversely ordered in C when comparing to that
in Fortran. Actually, I think it would result in more confusion if
we transpose the box array in advance for the user.